Operation Types fleshed out + Better op_type checking + Filter problems generated by possibility of execution
Did both of these tasks:

You added back the code that does type checking in PPG. The
restrictions on input-output types in the Operation classes are so
strict that no prediction problems are generated. Update the IOTYPES
class var for the ops.

Additionally, you changed labeler’s output, so labeler unit tests are
now failing. fix that.

Additionally, I noticed that there was some room for improvement in the
way the logic for param checking and setting is done so I renovated
that code.